pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-ai-assistant"
version = "1.0.0"
description = "Enterprise-grade internal Support AI Assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "opensearch-py>=2.4.0",
    "qdrant-client>=1.7.0",
    "openai>=1.12.0",
    "httpx>=0.26.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-prometheus>=0.43b0",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "tiktoken>=0.5.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
